+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes for the solr-cmd-utils project will be documented in this fi
 - add processing time filter
 - ElasticWriter - check if the index exists before creating a new one
 - ElasticWriter - implement delete for bulk methods
+- CSVReader - add a property to configure the quote
 
 
 ### Changed